requirements.txt
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.3